# WebAPI/TodoApp/BlazorOpenApiClient/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TodoAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/Authentication/token": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "GenerateToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Todos": {
      "get": {
        "tags": [
          "Todos"
        ],
        "operationId": "GetAllTodos",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Todos"
        ],
        "operationId": "CreateTodo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Todos/{todoId}": {
      "get": {
        "tags": [
          "Todos"
        ],
        "operationId": "GetOneTodo",
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TodoModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Todos"
        ],
        "operationId": "UpdateTodo",
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Todos"
        ],
        "operationId": "DeleteTodo",
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Todos/{todoId}/complete": {
      "put": {
        "tags": [
          "Todos"
        ],
        "operationId": "CompleteTodo",
        "parameters": [
          {
            "name": "todoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthenticationData": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TodoModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "task": {
            "type": "string",
            "nullable": true
          },
          "assignedTo": {
            "type": "integer",
            "format": "int32"
          },
          "isComplete": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Json Web Token"
      }
    }
  }
}